added update profile db logic
posgres RLS is not working for some reason. checking for auth.role() or
auth.id() won't work. the functions auth.session() & auth.user() returns
null, so the problem might be here. for now, ideal db rules will be in
the db/schema.sql file, but in the actual database no checks are done.